requirements.txt
pandas==2.1.0
requests==2.32.0
beautifulsoup4==4.12.2
openpyxl==3.1.2
streamlit==1.25.0
python-dotenv==1.0.1
